refactor: migrate loading overlay to ui components
Move ThemedLoadingOverlay from themed to ui components directory and
rename to loading-overlay.tsx. Update all import references and export
declarations to maintain component accessibility.
