+++ src/components/themed/index.ts
@@ -2,7 +2,6 @@ export * from './themed-button';
 export * from './themed-checkbox';
 export * from './themed-flatlist';
 export * from './themed-flashlist';
-export * from './themed-loading-overlay';
 export * from './themed-overlay';
 export * from './themed-radio';
 export * from './themed-safe-area-view';
