+++ src/components/ui/index.ts
@@ -3,4 +3,5 @@ export * from './avatar';
 export * from './confirm-overlay';
 export * from './icon-symbol';
 export * from './input-overlay';
+export * from './loading-overlay';
 export * from './selection-overlay';
